Refactor exceptions dir - property and missing entity exceptions into their own directories - move reusable types into a global dir
