+++ src/exception/property-exceptions/util.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2024 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,25 +17,25 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { RepoError } from './repo/types';
+import { RepoError } from './error-handling';
 import { EntityException, ProgramException, EntityValues, Entity } from './types';
 import _ from 'lodash';
 
 export function isProgramException(result: ProgramException | null): result is ProgramException {
-  return result?.programId !== undefined;
+	return result?.programId !== undefined;
 }
 
 export function isEntityException(result: EntityException | null): result is EntityException {
-  return (
-    result?.programId !== undefined &&
-    (result?.specimen !== undefined || result?.follow_up !== undefined)
-  );
+	return (
+		result?.programId !== undefined &&
+		(result?.specimen !== undefined || result?.follow_up !== undefined)
+	);
 }
 
 export function isRepoError(
-  result: ProgramException | EntityException | RepoError,
+	result: ProgramException | EntityException | RepoError,
 ): result is RepoError {
-  return result === RepoError.DOCUMENT_UNDEFINED || result === RepoError.SERVER_ERROR;
+	return result === RepoError.DOCUMENT_UNDEFINED || result === RepoError.SERVER_ERROR;
 }
 
 /**
@@ -44,7 +44,7 @@ export function isRepoError(
  * @returns snaked_cased string
  */
 export function normalizeEntityFileType(fileType: string) {
-  return _.snakeCase(fileType);
+	return _.snakeCase(fileType);
 }
 
 /**
@@ -53,5 +53,5 @@ export function normalizeEntityFileType(fileType: string) {
  * @returns true if valid Entity
  */
 export function isValidEntityType(fileType: string): fileType is Entity {
-  return Object.keys(EntityValues).includes(fileType);
+	return Object.keys(EntityValues).includes(fileType);
 }
